correction

+++ lib/libft/vector.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   vector.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kearmand <kearmand@student.42.fr>          +#+  +:+       +#+        */
+/*   By: ertrigna <ertrigna@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/21 16:24:14 by kearmand          #+#    #+#             */
-/*   Updated: 2025/05/06 09:42:46 by kearmand         ###   ########.fr       */
+/*   Updated: 2026/06/01 13:09:56 by ertrigna         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -63,7 +63,7 @@ int	vector_push_back(t_vector *vec, void *elem)
  */
 void vector_remove(t_vector *vec, size_t index)
 {
-	int i;
+	size_t i;
 
 	if (!vec || index >= vec->count)
 		return;
